50hztest2

+++ config.h
@@ -76,7 +76,7 @@ volatile uint8_t over_temp_counter = 0;
 #define MAXDELAY50 310 
 #define MINDELAY 35
 
-static uint8_t maxdelay = MAXDELAY;
+static uint16_t maxdelay = MAXDELAY;
 
 
 
@@ -108,7 +108,7 @@ uint8_t sleep_deviation = 13;
 float temp_sense = 0.0;
 uint8_t display_timer = 0;
 uint8_t idle_timeout = 60;
-uint8_t idle_decrease = 500;
+uint16_t idle_decrease = 500;
 
 float pid_integral = 0;
 float pid_last_error = 0;
